Rewrite stdio

+++ lib/libc/include/unistd.h
@@ -32,8 +32,8 @@ extern "C" {
 	int close(int fd);
 	char* getcwd(char* buf, size_t size);
 	off_t lseek(int fd, off_t offset, int whence);
-	int read(int fd, void* buf, size_t count);
-	int write(int fd, const void* buf, size_t count);
+	ssize_t read(int fd, void* buf, size_t count);
+	ssize_t write(int fd, const void* buf, size_t count);
 
 	// Process
 	void _exit(int status);
